+++ CHANGELOG.md
@@ -3,3 +3,12 @@
 ## 1.0.0 (2014-05-09)
 
 * Initial commit
+
+## 1.0.1 (2014-05-19)
+
+Upgrade Docker to 0.11 RC
+
+* Default to installing from the Docker repository instead of Canonical repo
+* Add ability to install from Docker repository, as Canonical repo lags
+* Giver Docker non-root access
+* Add docker alias when installed via docker.io package with update-alternatives